refactor: 메서드 네이밍 변경 (HttpResponseNew -> HttpResponse)
