+++ tomcat/src/main/java/org/apache/coyote/http11/Controller.java
@@ -2,5 +2,5 @@ package org.apache.coyote.http11;
 
 public interface Controller {
 
-    void service(HttpRequest request, HttpResponseNew response);
+    void service(HttpRequest request, HttpResponse response);
 }
