+++ tomcat/src/main/java/org/apache/coyote/http11/HttpResponse.java
@@ -1,21 +1,21 @@
 package org.apache.coyote.http11;
 
+import java.io.IOException;
+import java.io.OutputStream;
 import org.apache.catalina.session.Session;
 import org.apache.catalina.session.SessionManager;
 
 public class HttpResponse {
     private static final String CONTENT_TYPE_FORMAT = "%s;charset=utf-8";
 
+    private final OutputStream outputStream;
     private HttpResponseHeader headers;
     private HttpStatusCode statusCode;
-    private String body;
+    private String body; //TODO: Response에서 때어낼 수 있을 것 같은데?
 
-    private HttpResponse() {
+    public HttpResponse(OutputStream outputStream) {
+        this.outputStream = outputStream;
         this.headers = new HttpResponseHeader();
-    }
-
-    public static HttpResponse builder() {
-        return new HttpResponse();
     }
 
     public HttpResponse statusCode(HttpStatusCode statusCode) {
@@ -23,10 +23,18 @@ public class HttpResponse {
         return this;
     }
 
-    public HttpResponse responseBody(String body) {
+    public void responseBody(String body) {
         setDefaultHaders(HttpContentType.TEXT, body);
         this.body = body;
-        return this;
+        write();
+    }
+
+    private void write() {
+        try (outputStream) {
+            outputStream.write(build().getBytes());
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     private void setDefaultHaders(HttpContentType httpContentType, String body) {
@@ -34,19 +42,19 @@ public class HttpResponse {
         this.headers.addHeader("Content-Length", String.valueOf(body.getBytes().length));
     }
 
-    public HttpResponse staticResource(String path) {
+    public void staticResource(String path) {
         StaticResourceLoader loader = new StaticResourceLoader();
         String resource = loader.load(path);
         if (resource.isEmpty()) {
             String notFoundResource = loader.load("/404.html");
-            return this.statusCode(HttpStatusCode.NOT_FOUND)
+            this.statusCode(HttpStatusCode.NOT_FOUND)
                     .responseBody(notFoundResource);
         }
 
         HttpContentType contentType = HttpContentType.matchContentType(path);
         setDefaultHaders(contentType, resource);
         this.body = resource;
-        return this;
+        write();
     }
 
     public String build() {
@@ -58,11 +66,11 @@ public class HttpResponse {
         );
     }
 
-    public HttpResponse redirect(String uri) {
+    public void redirect(String uri) {
         this.body = "";
         headers.addHeader("Location", uri);
         setDefaultHaders(HttpContentType.TEXT, "");
-        return this;
+        write();
     }
 
     public HttpResponse createSession(String name, Object object) {
